Rename to make explicit the reason for splitting files: symbols deprecated/included since specific versions
