+++ gtk/gtk_deprecated_since_3_14.go.h
@@ -1,0 +1,30 @@
+
+static GtkStatusIcon *
+toGtkStatusIcon(void *p)
+{
+	return (GTK_STATUS_ICON(p));
+}
+
+/*
+ * deprecated since version 3.14
+ */
+static GtkAlignment *
+toGtkAlignment(void *p)
+{
+	return (GTK_ALIGNMENT(p));
+}
+
+static GtkArrow *
+toGtkArrow(void *p)
+{
+	return (GTK_ARROW(p));
+}
+
+static GtkMisc *
+toGtkMisc(void *p)
+{
+	return (GTK_MISC(p));
+}
+/*
+ * End deprecated since version 3.14
+ */
