+++ gtk/gtk_since_3_10.go.h
@@ -58,33 +58,4 @@ toGtkStackSwitcher(void *p)
 	return (GTK_STACK_SWITCHER(p));
 }
 
-
-static GtkStatusIcon *
-toGtkStatusIcon(void *p)
-{
-	return (GTK_STATUS_ICON(p));
-}
-
-/*
- * deprecated since version 3.14
- */
-static GtkAlignment *
-toGtkAlignment(void *p)
-{
-	return (GTK_ALIGNMENT(p));
-}
-
-static GtkArrow *
-toGtkArrow(void *p)
-{
-	return (GTK_ARROW(p));
-}
-
-static GtkMisc *
-toGtkMisc(void *p)
-{
-	return (GTK_MISC(p));
-}
-/*
- * End deprecated since version 3.14
- */
+
