+++ gtk/gtk_since_3_18.go
@@ -1,0 +1,5 @@
+// +build !gtk_3_6,!gtk_3_8,!gtk_3_10,!gtk_3_12,!gtk_3_14,!gtk_3_16
+
+// See: https://developer.gnome.org/gtk3/3.18/api-index-3-18.html
+
+package gtk
