+++ gtk/gtk_since_3_8.go
@@ -14,16 +14,34 @@
 // ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
 // OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
-// This file includes wrapers for symbols included since GTK 3.12, and
+// This file includes wrapers for symbols included since GTK 3.8, and
 // and should not be included in a build intended to target any older GTK
-// versions.  To target an older build, such as 3.10, use
-// 'go build -tags gtk_3_10'.  Otherwise, if no build tags are used, GTK 3.12
+// versions.  To target an older build, such as 3.8, use
+// 'go build -tags gtk_3_8'.  Otherwise, if no build tags are used, GTK 3.18
 // is assumed and this file is built.
-// +build !gtk_3_6,!gtk_3_8,!gtk_3_10
+// +build !gtk_3_6
 
 package gtk
 
 // #cgo pkg-config: gtk+-3.0
-// #include <stdlib.h>
 // #include <gtk/gtk.h>
-// #include "gtk_3_12.go.h"
+import "C"
+
+/*
+ * Constants
+ */
+
+const (
+	STATE_FLAG_DIR_LTR StateFlags = C.GTK_STATE_FLAG_DIR_LTR
+	STATE_FLAG_DIR_RTL StateFlags = C.GTK_STATE_FLAG_DIR_RTL
+)
+
+/*
+ * GtkWidget
+ */
+
+// IsVisible is a wrapper around gtk_widget_is_visible().
+func (v *Widget) IsVisible() bool {
+	c := C.gtk_widget_is_visible(v.native())
+	return gobool(c)
+}
